More float16

+++ src/traintestgyms/ignitegym.py
@@ -1,3 +1,4 @@
+import torch
 import numpy as np
 from ignite.engine import Engine
 from ignite.engine import Events
@@ -64,7 +65,7 @@ class TrainTestGym:
 
     ### Training step
     def __train_step(self, engine, batch):
-        batch = batch.to(self.device)
+        batch = batch.to(self.device, dtype=torch.float16)
 
         if engine.t_start != 0:
             engine.t_start = batch[0,self.time_column_idx].item()
